Splits out usePagination into own files and updates the specs
