+++ src/components/Grid/Pagination.test.js
@@ -1,28 +1,18 @@
 import React from 'react'
 import TestRenderer, { act } from 'react-test-renderer'
 
-import Pagination, { usePagination } from './Pagination'
+import Pagination from './Pagination'
+import styles from './Pagination.module.css'
 
 const responseStub = {
   page: 1,
   per_page: 2,
   total: 6,
   total_pages: 3,
-  data: [
-    {
-      id: 1,
-      name: 'cerulean',
-      year: 2000,
-    },
-    {
-      id: 2,
-      name: 'fuchsia rose',
-      year: 2001,
-    },
-  ],
+  data: [{ id: 1 }, { id: 2 }],
 }
 
-describe('API', () => {
+describe('Pagination Component', () => {
   let tree
   let fetchFn
   let renderFn
@@ -75,9 +65,19 @@ describe('API', () => {
     expect(buttons.length).toEqual(5)
     const first = buttons[0]
     expect(first.props['aria-label']).toEqual('Goto Page 1')
-    expect(first.children).toEqual(['«'])
     const last = buttons[4]
     expect(last.props['aria-label']).toEqual('Goto Page 3')
-    expect(last.children).toEqual(['»'])
+  })
+
+  it('active page reflect current page', async () => {
+    await act(async () => {
+      tree = TestRenderer.create(
+        <Pagination fetchPageCallback={fetchFn} renderCallback={renderFn} />
+      )
+    })
+    const root = tree.root
+    const buttons = root.findAllByType('button')
+    const firstPageNumberButton = buttons[1]
+    expect(firstPageNumberButton.props.className).toContain(styles.active)
   })
 })
